Remove baseKeybase
